@@ -32,37 +32,34 @@
     "@babel/plugin-transform-runtime": "^7.18.2",
     "@babel/preset-env": "^7.18.2",
     "@babel/preset-react": "^7.17.12",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.7",
     "babel-loader": "^8.2.5",
     "clean-webpack-plugin": "^4.0.0",
     "css-loader": "^6.7.1",
-    "enzyme": "^3.11.0",
-    "enzyme-to-json": "^3.6.2",
-    "eslint": "^8.16.0",
+    "eslint": "^8.17.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-react": "^7.28.0",
-    "html-webpack-plugin": "^5.4.0",
-    "jest": "^27.4.7",
-    "jest-cli": "^27.4.7",
-    "jest-environment-jsdom": "^27.2.5",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-react": "^7.30.0",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^28.1.1",
+    "jest-cli": "^28.1.1",
+    "jest-environment-jsdom": "^28.1.1",
     "jest-transform-stub": "^2.0.0",
     "mini-css-extract-plugin": "^2.6.0",
-    "react-test-renderer": "^17.0.2",
+    "react-test-renderer": "^18.1.0",
     "redux-mock-store": "^1.5.4",
-    "sass": "^1.52.1",
+    "sass": "^1.52.3",
     "sass-loader": "^13.0.0",
     "style-loader": "^3.3.1",
-    "webpack": "^5.72.1",
+    "webpack": "^5.73.0",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.9.0"
+    "webpack-dev-server": "^4.9.2"
   },
   "dependencies": {
     "prop-types": "^15.8.1",
     "query-string": "^7.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-router": "^6.2.1",
     "react-router-dom": "^6.3.0"
   }
