@@ -1,15 +1,19 @@
 import React from 'react';
+import TestRenderer from 'react-test-renderer';
 import App from './App';
 
-const wrapper = shallow(<App />);
+const wrapper = TestRenderer.create(<App />);
+const instance = wrapper.root;
+
 describe('(Component) App', () => {
-
-  it('match the snapshot', () => {
-    const wrapper = renderer.create(<App />).toJSON();
-    expect(wrapper).toMatchSnapshot();
+  it('should match the snapshot', () => {
+    const tree = renderer
+      .create(<App />)
+      .toJSON();
+    expect(tree).toMatchSnapshot();
   });
-
+  
   it('should inject the Router component', () => {
-    expect(wrapper.find('.app').length).toBe(1);
+    expect(instance.findByProps({className: 'app'}).children.length).toBe(1);
   });
 });
