@@ -1,5 +1,9 @@
 import React from 'react';
 
-const Home = () => <div><h1>Hello World</h1></div>;
+const Home = () => (
+  <div>
+    <h1>Hello World</h1>
+  </div>
+);
 
 export default Home;
